api/requirements.txt
Django==2.0.4
djangorestframework==3.8.1
gunicorn==19.7.1
model-mommy==1.5.1
ipdb==0.11
django-rest-auth==0.9.3
django-allauth==0.35.0
django-cors-headers==2.2.0
coreapi==2.3.3
whitenoise==3.3.1
psycopg2==2.7.4
